begin making predicate forrest a bit more coherent
