+++ nomenklatura/matching/compare/identifiers.py
@@ -1,0 +1,7 @@
+from followthemoney.proxy import E
+
+
+# def imo_match(left: E, right: E) -> float:
+#     """Matching IMO numbers between the two entities."""
+#     lv, rv = type_pair(left, right, registry.imo)
+#     return has_overlap(set(lv), set(rv))
