+++ nomenklatura/matching/compare/util.py
@@ -1,0 +1,12 @@
+from typing import List, Set, Union
+
+
+def is_disjoint(
+    left: Union[Set[str], List[str]],
+    right: Union[Set[str], List[str]],
+) -> bool:
+    """Returns 1.0 if both sequences are non-empty but have no common values."""
+    if len(left) and len(right):
+        if set(left).isdisjoint(right):
+            return True
+    return False
